Only do this on Sunday

+++ build_nightly.sh
@@ -4,7 +4,10 @@ while true
 do
 	LIFERAY_DOCKER_IMAGE_FILTER=7.4.13.nightly ./build_all_images.sh --push
 
-	docker system prune --all --force
+	if [ $(date +%w) == 0 ]
+	then
+		docker system prune --all --force
+	fi
 
 	echo ""
 	echo `date`
